Move AverageRecordRate to core module
